+++ Src/AwesomeAssertions/Primitives/StringAssertions.cs
@@ -194,16 +194,8 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         string unexpected,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            BeEquivalentTo(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(notEquivalent)
+            .ForFailingAssertion(() => BeEquivalentTo(unexpected))
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:string} not to be equivalent to {0}{reason}, but they are.", unexpected);
 
@@ -234,16 +226,8 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().BeEquivalentTo(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(notEquivalent)
+            .ForFailingAssertion(() => BeEquivalentTo(unexpected, config))
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:string} not to be equivalent to {0}{reason}, but they are.", unexpected);
 
@@ -1084,18 +1068,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWith(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to start with {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .ForFailingAssertion(() => StartWith(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1186,19 +1166,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWithEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Expected {context:string} not to start with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .ForFailingAssertion(() => StartWithEquivalentOf(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1227,19 +1202,15 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWithEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Expected {context:string} not to start with equivalent of {0}{reason}, but found <null>.",
+                unexpected, chain => chain
+                    .ForCondition(Subject is not null)
+                    .FailWith(", but found <null>.")
+                    .Then
+                    .ForFailingAssertion(() => StartWithEquivalentOf(unexpected, config))
+                    .FailWith(", but found {0}.", unexpected, Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1293,18 +1264,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWith(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .ForFailingAssertion(() => Subject.Should().EndWith(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1395,18 +1362,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWithEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with equivalent of {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .ForFailingAssertion(() => EndWithEquivalentOf(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1435,18 +1398,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWithEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with equivalent of {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .ForFailingAssertion(() => EndWithEquivalentOf(unexpected, config))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1884,24 +1843,13 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         assertionChain
-            .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
-
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().ContainEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
-        assertionChain
-            .ForCondition(notEquivalent)
-            .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason} but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Did not expect {context:string} to contain the equivalent of {0}{reason}", unexpected, chain => chain
+                .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
+                .FailWith(", but found {0}.", Subject)
+                .Then
+                .ForFailingAssertion(() => ContainEquivalentOf(unexpected))
+                .FailWith(", but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1929,24 +1877,13 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(config);
 
         assertionChain
-            .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
-
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().ContainEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
-        assertionChain
-            .ForCondition(notEquivalent)
-            .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Did not expect {context:string} to contain the equivalent of {0}{reason}", unexpected, chain => chain
+                .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
+                .FailWith(", but found {0}.", Subject)
+                .Then
+                .ForFailingAssertion(() => ContainEquivalentOf(unexpected, config))
+                .FailWith(", but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }